+++ lib/controllers/migrate-controller.ts
@@ -122,7 +122,7 @@ export class MigrateController
 		{
 			packageName: "@nativescript/core",
 			minVersion: "6.5.0",
-			desiredVersion: "~8.7.0",
+			desiredVersion: "~8.8.0",
 			shouldAddIfMissing: true,
 		},
 		{
@@ -132,7 +132,7 @@ export class MigrateController
 		{
 			packageName: "@nativescript/types",
 			minVersion: "7.0.0",
-			desiredVersion: "~8.7.0",
+			desiredVersion: "~8.8.0",
 			isDev: true,
 		},
 		{
@@ -191,7 +191,7 @@ export class MigrateController
 		{
 			packageName: "@nativescript/angular",
 			minVersion: "10.0.0",
-			desiredVersion: "^17.0.0",
+			desiredVersion: "^18.0.0",
 			async shouldMigrateAction(
 				dependency: IMigrationDependency,
 				projectData: IProjectData,
@@ -296,13 +296,13 @@ export class MigrateController
 		{
 			packageName: "@nativescript/ios",
 			minVersion: "6.5.3",
-			desiredVersion: "~8.7.0",
+			desiredVersion: "~8.8.0",
 			isDev: true,
 		},
 		{
 			packageName: "@nativescript/android",
 			minVersion: "7.0.0",
-			desiredVersion: "~8.7.0",
+			desiredVersion: "~8.8.0",
 			isDev: true,
 		},
 	];
@@ -1311,7 +1311,7 @@ export class MigrateController
 
 	private async migrateNativeScriptAngular(): Promise<IMigrationDependency[]> {
 		const minVersion = "10.0.0";
-		const desiredVersion = "~17.3.0";
+		const desiredVersion = "~18.0.0";
 
 		const dependencies: IMigrationDependency[] = [
 			{