+++ lib/declarations.d.ts
@@ -578,13 +578,19 @@ interface IAndroidBundleOptions {
 	aab: boolean;
 }
 
-interface IAndroidOptions {
-	gradleArgs: string[];
-	gradleFlavor: string;
+interface IEmbedOptions {
+	hostProjectPath: string;
+	hostProjectModuleName: string;
+}
+
+interface IAndroidOptions extends IEmbedOptions {
 	gradlePath: string;
+	gradleFlavor: string;
 	gradleArgs: string[];
 	overrideRuntimeGradleFiles: boolean;
 }
+
+interface IIOSOptions extends IEmbedOptions {}
 
 interface ITypingsOptions {
 	jar: string;
@@ -606,6 +612,7 @@ interface IOptions
 		IProvision,
 		ITeamIdentifier,
 		IAndroidOptions,
+		IIOSOptions,
 		IAndroidReleaseOptions,
 		IAndroidBundleOptions,
 		INpmInstallConfigurationOptions,
@@ -704,6 +711,9 @@ interface IOptions
 	dryRun: boolean;
 
 	platformOverride: string;
+
+	// allow arbitrary options
+	[optionName: string]: any;
 }
 
 interface IEnvOptions {
@@ -1012,6 +1022,7 @@ interface IXcprojService {
 	 * @return {string} The full path to the xcodeproj
 	 */
 	getXcodeprojPath(projectData: IProjectData, projectRoot: string): string;
+	findXcodeProject(dir: string): string;
 }
 
 /**