+++ lib/definitions/project.d.ts
@@ -173,6 +173,7 @@ interface INsConfig {
 	visionos?: INSConfigVisionOS;
 	ignoredNativeDependencies?: string[];
 	hooks?: INsConfigHooks[];
+	projectName?: string;
 }
 
 interface IProjectData extends ICreateProjectData {
@@ -207,6 +208,7 @@ interface IProjectData extends ICreateProjectData {
 	 * The value can be changed by setting `webpackConfigPath` in nativescript.config.
 	 */
 	webpackConfigPath: string;
+	projectName: string;
 
 	/**
 	 * Initializes project data with the given project directory. If none supplied defaults to --path option or cwd.