+++ lib/options.ts
@@ -10,6 +10,7 @@ import {
 	ISettingsService,
 } from "./common/declarations";
 import { injector } from "./common/yok";
+import { APP_FOLDER_NAME } from "./constants";
 export class Options {
 	private static DASHED_OPTION_REGEX = /(.+?)([A-Z])(.*)/;
 	private static NONDASHED_OPTION_REGEX = /(.+?)[-]([a-zA-Z])(.*)/;
@@ -226,13 +227,20 @@ export class Options {
 				hasSensitiveValue: false,
 			},
 			overrideRuntimeGradleFiles: { type: OptionType.Boolean, hasSensitiveValue: false },
+			
+			gradleFlavor: { type: OptionType.String, hasSensitiveValue: false },
+			gradlePath: { type: OptionType.String, hasSensitiveValue: false },
 			gradleArgs: {
 				type: OptionType.String,
 				hasSensitiveValue: false,
 				array: true,
 			},
-			gradleFlavor: { type: OptionType.String, hasSensitiveValue: false },
-			gradlePath: { type: OptionType.String, hasSensitiveValue: false },
+			hostProjectPath: { type: OptionType.String, hasSensitiveValue: false },
+			hostProjectModuleName: {
+				type: OptionType.String,
+				hasSensitiveValue: false,
+				default: APP_FOLDER_NAME,
+			},
 			aab: { type: OptionType.Boolean, hasSensitiveValue: false },
 			filterDevicesArch: { type: OptionType.Boolean, hasSensitiveValue: false },
 			performance: { type: OptionType.Object, hasSensitiveValue: true },