+++ lib/services/android-project-service.ts
@@ -169,28 +169,30 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 			);
 		}
 		if (projectData && projectData.platformsDir) {
-			const projectRoot = path.join(
-				projectData.platformsDir,
-				AndroidProjectService.ANDROID_PLATFORM_NAME
-			);
+			const projectRoot = this.$options.hostProjectPath
+				? this.$options.hostProjectPath
+				: path.join(
+						projectData.platformsDir,
+						AndroidProjectService.ANDROID_PLATFORM_NAME
+				  );
 
 			const appDestinationDirectoryArr = [
 				projectRoot,
-				constants.APP_FOLDER_NAME,
+				this.$options.hostProjectModuleName,
 				constants.SRC_DIR,
 				constants.MAIN_DIR,
 				constants.ASSETS_DIR,
 			];
 			const configurationsDirectoryArr = [
 				projectRoot,
-				constants.APP_FOLDER_NAME,
+				this.$options.hostProjectModuleName,
 				constants.SRC_DIR,
 				constants.MAIN_DIR,
 				constants.MANIFEST_FILE_NAME,
 			];
 			const deviceBuildOutputArr = [
 				projectRoot,
-				constants.APP_FOLDER_NAME,
+				this.$options.hostProjectModuleName,
 				constants.BUILD_DIR,
 				constants.OUTPUTS_DIR,
 				constants.APK_DIR,
@@ -213,7 +215,7 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 					if (buildOptions.androidBundle) {
 						return path.join(
 							projectRoot,
-							constants.APP_FOLDER_NAME,
+							this.$options.hostProjectModuleName,
 							constants.BUILD_DIR,
 							constants.OUTPUTS_DIR,
 							constants.BUNDLE_DIR
@@ -232,8 +234,8 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 					if (buildOptions.androidBundle) {
 						return {
 							packageNames: [
-								`${constants.APP_FOLDER_NAME}${constants.AAB_EXTENSION_NAME}`,
-								`${constants.APP_FOLDER_NAME}-${buildMode}${constants.AAB_EXTENSION_NAME}`,
+								`${this.$options.hostProjectModuleName}${constants.AAB_EXTENSION_NAME}`,
+								`${this.$options.hostProjectModuleName}-${buildMode}${constants.AAB_EXTENSION_NAME}`,
 							],
 						};
 					}
@@ -243,11 +245,11 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 							`${packageName}-${buildMode}${constants.APK_EXTENSION_NAME}`,
 							`${projectData.projectName}-${buildMode}${constants.APK_EXTENSION_NAME}`,
 							`${projectData.projectName}${constants.APK_EXTENSION_NAME}`,
-							`${constants.APP_FOLDER_NAME}-${buildMode}${constants.APK_EXTENSION_NAME}`,
+							`${this.$options.hostProjectModuleName}-${buildMode}${constants.APK_EXTENSION_NAME}`,
 						],
 						regexes: [
 							new RegExp(
-								`(${packageName}|${constants.APP_FOLDER_NAME})-.*-(${Configurations.Debug}|${Configurations.Release})(-unsigned)?${constants.APK_EXTENSION_NAME}`,
+								`(${packageName}|${this.$options.hostProjectModuleName})-.*-(${Configurations.Debug}|${Configurations.Release})(-unsigned)?${constants.APK_EXTENSION_NAME}`,
 								"i"
 							),
 						],
@@ -634,7 +636,7 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 		return this.$fs.exists(
 			path.join(
 				this.getPlatformData(projectData).appDestinationDirectoryPath,
-				constants.APP_FOLDER_NAME
+				this.$options.hostProjectModuleName
 			)
 		);
 	}
@@ -825,10 +827,12 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 		projectData: IProjectData,
 		dependencies: IDependencyData[]
 	): IDependencyData[] {
-		const platformDir = path.join(
-			projectData.platformsDir,
-			AndroidProjectService.ANDROID_PLATFORM_NAME
-		);
+		const platformDir = this.$options.hostProjectPath
+			? this.$options.hostProjectPath
+			: path.join(
+					projectData.platformsDir,
+					AndroidProjectService.ANDROID_PLATFORM_NAME
+			  );
 		const dependenciesJsonPath = path.join(
 			platformDir,
 			constants.DEPENDENCIES_JSON_NAME
@@ -995,7 +999,7 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 		projectData: IProjectData
 	): string {
 		const resourcePath: string[] = [
-			constants.APP_FOLDER_NAME,
+			this.$options.hostProjectModuleName,
 			constants.SRC_DIR,
 			constants.MAIN_DIR,
 			constants.RESOURCES_DIR,
@@ -1011,7 +1015,7 @@ export class AndroidProjectService extends projectServiceBaseLib.PlatformProject
 		projectData: IProjectData
 	): string {
 		const resourcePath: string[] = [
-			constants.APP_FOLDER_NAME,
+			this.$options.hostProjectModuleName,
 			constants.SRC_DIR,
 		];