+++ lib/services/plugins-service.ts
@@ -806,7 +806,7 @@ This framework comes from ${dependencyName} plugin, which is installed multiple 
 				);
 				const pluginDestinationPath = path.join(
 					platformData.appDestinationDirectoryPath,
-					constants.APP_FOLDER_NAME,
+					this.$options.hostProjectModuleName,
 					"tns_modules"
 				);
 				await action(
@@ -879,6 +879,10 @@ This framework comes from ${dependencyName} plugin, which is installed multiple 
 	private getAllPluginsNativeHashes(
 		pathToPluginsBuildFile: string
 	): IDictionary<IStringDictionary> {
+		if (this.$options.hostProjectPath) {
+			// TODO: force rebuild plugins for now until we decide where to put .ns-plugins-build-data.json when embedding
+			return {};
+		}
 		let data: IDictionary<IStringDictionary> = {};
 		if (this.$fs.exists(pathToPluginsBuildFile)) {
 			data = this.$fs.readJson(pathToPluginsBuildFile);
@@ -893,6 +897,11 @@ This framework comes from ${dependencyName} plugin, which is installed multiple 
 		currentPluginNativeHashes: IStringDictionary;
 		allPluginsNativeHashes: IDictionary<IStringDictionary>;
 	}): void {
+		if (this.$options.hostProjectPath) {
+			// TODO: force rebuild plugins for now until we decide where to put .ns-plugins-build-data.json when embedding
+			return;
+		}
+
 		opts.allPluginsNativeHashes[opts.pluginData.name] =
 			opts.currentPluginNativeHashes;
 		this.$fs.writeJson(