+++ lib/services/prepare-data-service.ts
@@ -1,10 +1,11 @@
 import { IOSPrepareData, AndroidPrepareData } from "../data/prepare-data";
 import { injector } from "../common/yok";
+import { IOptions } from "../declarations";
 
 export class PrepareDataService implements IPrepareDataService {
 	constructor(private $mobileHelper: Mobile.IMobileHelper) {}
 
-	public getPrepareData(projectDir: string, platform: string, data: any) {
+	public getPrepareData(projectDir: string, platform: string, data: IOptions) {
 		const platformLowerCase = platform.toLowerCase();
 
 		if (this.$mobileHelper.isApplePlatform(platform)) {