+++ lib/services/project-changes-service.ts
@@ -23,6 +23,7 @@ import {
 } from "../definitions/project-changes";
 import * as _ from "lodash";
 import { injector } from "../common/yok";
+import { IOptions } from "../declarations";
 
 const prepareInfoFileName = ".nsprepareinfo";
 
@@ -62,6 +63,7 @@ export class ProjectChangesService implements IProjectChangesService {
 		private $devicePlatformsConstants: Mobile.IDevicePlatformsConstants,
 		private $fs: IFileSystem,
 		private $logger: ILogger,
+		private $options: IOptions,
 		public $hooksService: IHooksService,
 		private $nodeModulesDependenciesBuilder: INodeModulesDependenciesBuilder
 	) {}
@@ -236,6 +238,11 @@ export class ProjectChangesService implements IProjectChangesService {
 	}
 
 	public getPrepareInfo(platformData: IPlatformData): IPrepareInfo {
+		if (this.$options.hostProjectPath) {
+			// TODO: always prepare for now until we decide where to keep the .nsprepareinfo file when embedding
+			return null;
+		}
+
 		const prepareInfoFilePath = this.getPrepareInfoFilePath(platformData);
 		let prepareInfo: IPrepareInfo = null;
 		if (this.$fs.exists(prepareInfoFilePath)) {
@@ -256,6 +263,11 @@ export class ProjectChangesService implements IProjectChangesService {
 	): Promise<void> {
 		if (!this._prepareInfo) {
 			await this.ensurePrepareInfo(platformData, projectData, prepareData);
+		}
+
+		if (this.$options.hostProjectPath) {
+			// TODO: do not save for now until we decide where to keep the .nsprepareinfo file when embedding
+			return null;
 		}
 
 		const prepareInfoFilePath = this.getPrepareInfoFilePath(platformData);